auth requests

+++ app/src/components/WorkoutDetails.js
@@ -1,14 +1,23 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
-  const handleDelete = async () => {
+  const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
@@ -29,11 +38,9 @@ const WorkoutDetails = ({ workout }) => {
         {workout.reps}
       </p>
       <p>
-        {formatDistanceToNow(new Date(workout.createdAt), {
-          addSuffix: true,
-        })}
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
       </p>
-      <span className="material-symbols-outlined" onClick={handleDelete}>
+      <span className="material-symbols-outlined" onClick={handleClick}>
         delete
       </span>
     </div>
